Fixed end_round and added coinSearch and swordSearch
end_round did not swap current and other, leading for both a and b being the same and resulting in only ties

coinSearch and swordSearch finds x card in deck and adds it to the player's and opponent's hands respectively. swords might be never used as devil card exists but whatever.